+++ planning/grid_base/bfs/CMakeLists.txt
@@ -1,4 +1,4 @@
-cmake_minimum_required(VERSION 3.15)
+cmake_minimum_required(VERSION 3.22)
 
 include_directories(
     ${CMAKE_CURRENT_SOURCE_DIR}
@@ -10,11 +10,19 @@ link_directories(
     ${PROJECT_SOURCE_DIR}
 )
 
-add_library(
-    bfs SHARED
-    bfs.cpp
-    bfs.h
-    ${PROJECT_SOURCE_DIR}/planning/i_planning.h
-    ${PROJECT_SOURCE_DIR}/planning/grid_base/common_grid_base.h
+set(
+    BFS_FILES
+    ${CMAKE_CURRENT_SOURCE_DIR}/bfs.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/bfs.h
 )
 
+add_library(
+    bfs SHARED
+    ${BFS_FILES}
+)
+
+target_link_libraries(
+    bfs
+    PUBLIC
+    common_grid_base
+)
