+++ planning/tree_base/rrt/CMakeLists.txt
@@ -1,4 +1,4 @@
-cmake_minimum_required(VERSION 3.15)
+cmake_minimum_required(VERSION 3.22)
 
 include_directories(
     ${CMAKE_CURRENT_SOURCE_DIR}
