+++ test/test_access_data.cpp
@@ -9,7 +9,7 @@
  *
  */
 
-#include "planning/common_planning.h"
+#include "planning/grid_base/include/common_grid_base.h"
 
 #include <fstream>
 #include <gtest/gtest.h>
