+++ test/test_map.cpp
@@ -9,7 +9,7 @@
  *
  */
 
-#include "planning/common_planning.h"
+#include "planning/grid_base/include/common_grid_base.h"
 #include "test_fixture.h"
 #include <gtest/gtest.h>
